Move ThemeContext to the serivces folder
