+++ src/actions/WalletListMenuActions.js
@@ -9,12 +9,12 @@ import { sprintf } from 'sprintf-js'
 import { launchModal } from '../components/common/ModalProvider.js'
 import { RawTextModal } from '../components/modals/RawTextModal.js'
 import { Airship, showError } from '../components/services/AirshipInstance.js'
+import { getTheme } from '../components/services/ThemeContext.js'
 import * as Constants from '../constants/indexConstants'
 import s from '../locales/strings.js'
 import Text from '../modules/UI/components/FormattedText/FormattedText.ui.js'
 import * as WALLET_SELECTORS from '../modules/UI/selectors.js'
 import { B } from '../styles/common/textStyles.js'
-import { getTheme } from '../theme/ThemeContext.js'
 import type { Dispatch, GetState } from '../types/reduxTypes.js'
 import { getWalletName } from '../util/CurrencyWalletHelpers.js'
 import { showDeleteWalletModal } from './DeleteWalletModalActions.js'
