+++ src/components/common/SettingsRow.js
@@ -3,7 +3,7 @@
 import * as React from 'react'
 import { Text, TouchableHighlight, View } from 'react-native'
 
-import { type Theme, type ThemeProps, cacheStyles, withTheme } from '../../theme/ThemeContext.js'
+import { type Theme, type ThemeProps, cacheStyles, withTheme } from '../services/ThemeContext.js'
 
 type OwnProps = {
   disabled?: boolean, // Show with grey style
