+++ src/components/modals/TransactionAdvanceDetails.js
@@ -6,8 +6,8 @@ import SafariView from 'react-native-safari-view'
 import Entypo from 'react-native-vector-icons/Entypo'
 
 import s from '../../locales/strings.js'
-import { type Theme, type ThemeProps, cacheStyles, withTheme } from '../../theme/ThemeContext.js'
 import { Tile } from '../common/Tile.js'
+import { type Theme, type ThemeProps, cacheStyles, withTheme } from '../services/ThemeContext.js'
 import { ThemedModal } from '../themed/ThemedModal.js'
 import { type AirshipBridge } from './modalParts'
