+++ src/components/scenes/PromotionSettingsScene.js
@@ -9,7 +9,6 @@ import { sprintf } from 'sprintf-js'
 
 import { activatePromotion, removePromotion } from '../../actions/AccountReferralActions.js'
 import s from '../../locales/strings.js'
-import { type Theme, type ThemeProps, cacheStyles, withTheme } from '../../theme/ThemeContext.js'
 import { THEME } from '../../theme/variables/airbitz.js'
 import { type Dispatch, type State as ReduxState } from '../../types/reduxTypes.js'
 import { type AccountReferral, type DeviceReferral } from '../../types/ReferralTypes.js'
@@ -18,6 +17,7 @@ import { SceneWrapper } from '../common/SceneWrapper.js'
 import { SettingsHeaderRow } from '../common/SettingsHeaderRow.js'
 import { SettingsRow } from '../common/SettingsRow.js'
 import { showActivity } from '../services/AirshipInstance.js'
+import { type Theme, type ThemeProps, cacheStyles, withTheme } from '../services/ThemeContext.js'
 
 type StateProps = {
   accountReferral: AccountReferral,
