+++ src/modules/UI/components/Gradient/Gradient.ui.js
@@ -4,7 +4,7 @@ import * as React from 'react'
 import { StyleSheet } from 'react-native'
 import LinearGradient from 'react-native-linear-gradient'
 
-import { type ThemeProps, withTheme } from '../../../../theme/ThemeContext.js'
+import { type ThemeProps, withTheme } from '../../../../components/services/ThemeContext.js'
 
 const UPPER_LEFT = { x: 0, y: 0 }
 const UPPER_RIGHT = { x: 1, y: 0 }
